Replace RFC 6844 with RFC 8659.
This also removes an addendum since the amended behavior is now part of
the core spec. It removes a clarification about Relevant RRSets
that don't contain issue or issuewild, but also aren't empty (by
definition), since that clarification has been incorporated into the
core spec.

Since a lower bound on CNAME following was part of the addendum, but
didn't make it into RFC 8659, this change hoists that requirement out of
the addendum and into the overall CAA requirement section.

Since RFC 8659 codifies the current behavior and clarifies some syntax,
this does not have a phase-in period. CAA implementations that were
compliant with RFC 6844 as amended in the BRs will also be compliant
with RFC 8659.